+++ resources/model_proof_params_attributes.go
@@ -13,9 +13,9 @@ type ProofParamsAttributes struct {
 	EventId                   string  `json:"event_id"`
 	ExpirationDateLowerBound  string  `json:"expiration_date_lower_bound"`
 	ExpirationDateUpperBound  string  `json:"expiration_date_upper_bound"`
-	IdentityCounter           int32   `json:"identity_counter"`
-	IdentityCounterLowerBound int32   `json:"identity_counter_lower_bound"`
-	IdentityCounterUpperBound int32   `json:"identity_counter_upper_bound"`
+	IdentityCounter           int64   `json:"identity_counter"`
+	IdentityCounterLowerBound int64   `json:"identity_counter_lower_bound"`
+	IdentityCounterUpperBound int64   `json:"identity_counter_upper_bound"`
 	Selector                  string  `json:"selector"`
 	TimestampLowerBound       string  `json:"timestamp_lower_bound"`
 	TimestampUpperBound       string  `json:"timestamp_upper_bound"`
